refactored BankAccount into interface and updated SavingAccount and CurrentAccount updated test depositWithPositiveAmountTest() to check if transaction is added to the list of transactions
