+++ src/main/java/com/booleanuk/core/BankAccount.java
@@ -1,28 +1,21 @@
 package com.booleanuk.core;
 
-public class BankAccount {
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
 
-    private double accountBalance;
-    public BankAccount(){
-        accountBalance =0.00;
+public interface BankAccount {
 
-    }
 
-    public boolean deposit(double amount) {
-        if(amount > 0) {
-            accountBalance += amount;
-            return true;
-        }else {
-            System.out.println("Amount must be a positive number");
-            return false;
-        }
-    }
 
-    public boolean withdraw(double amount) {
-        return true;
-    }
 
-    public double getAccountBalance() {
-        return accountBalance;
-    }
+
+
+
+
+    boolean deposit(double amount);
+
+    public boolean withdraw(double amount);
+
+
 }
