+++ src/main/java/com/booleanuk/core/SavingAccount.java
@@ -1,4 +1,13 @@
 package com.booleanuk.core;
 
-public class SavingAccount extends BankAccount{
+public class SavingAccount implements BankAccount{
+    @Override
+    public boolean deposit(double amount) {
+        return false;
+    }
+
+    @Override
+    public boolean withdraw(double amount) {
+        return false;
+    }
 }
